+++ src/main/java/com/exam/useful/usecase/user/ERole.java
@@ -1,4 +1,4 @@
-package com.exam.useful.service.user;
+package com.exam.useful.usecase.user;
 
 public enum ERole {
     ROLE_ADMIN, ROLE_MANAGER, ROLE_USER
